% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A1-weights_med.R, R/A2-weights_ipw.R,
%   R/A3-weights_cmodds.R, R/C3a-estimate_Ypred.R
\name{.check_plot.med}
\alias{.check_plot.med}
\alias{.check_plot.ipw}
\alias{.check_plot.cm.odds}
\alias{.check_plot.Ypred}
\alias{dot-check_plot}
\title{Check inputs for plotting}
\usage{
.check_plot.med(env)

.check_plot.ipw(env)

.check_plot.cm.odds(env)

.check_plot.Ypred(env)
}
\arguments{
\item{env}{An environment}
}
\description{
Internal function to check plot inputs within functions \code{get_weights_med} and \code{estimate_wtd}
}
\details{
Checks if c.vars.std and m.vars.std are not NULL (ie specified by user) and are part of c.vars and m.vars respectively.

Check if variables named in c.vars.std and m.vars.std in data can be standardized (ie are continuous variables).
}
