% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B1-generic_dispatch.R
\name{.clean_estimator}
\alias{.clean_estimator}
\title{Clean typos in estimator name}
\usage{
.clean_estimator(estimator)
}
\arguments{
\item{estimator}{The estimator name as typed by user.}
}
\value{
The right estimator name (correcting anticipated typos).
}
\description{
Clean typos in estimator name
}
\keyword{internal}
