% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C5a-estimate_MsimYpred.R,
%   R/C5b-estimate_MsimYpredMR.R
\name{.clean_m.MsimYpred}
\alias{.clean_m.MsimYpred}
\alias{.clean_m.MsimYpredMR}
\alias{dot-clean_m}
\title{Internal: clean inputs related to mediator model(s)}
\usage{
.clean_m.MsimYpred(env)

.clean_m.MsimYpredMR(env)
}
\arguments{
\item{env}{An environment}
}
\description{
Internal: clean inputs related to mediator model(s)
}
\keyword{internal}
