% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A1-weights_med.R, R/A2-weights_ipw.R,
%   R/A3-weights_cmodds.R, R/C1a-estimate_Y2pred.R
\name{.clean_weights.med}
\alias{.clean_weights.med}
\alias{.clean_weights.ipw}
\alias{.clean_weights.cm.odds}
\alias{.clean_a.var}
\alias{dot-clean_weights}
\title{Clean inputs for weighting}
\usage{
.clean_weights.med(env)

.clean_weights.ipw(env)

.clean_weights.cm.odds(env)

.clean_a.var(env)
}
\arguments{
\item{env}{An environment}
}
\description{
Internal functions called by \code{get_weights.} or \code{estimate_} functions to clean the inputs used for the weighting method such as \code{max.stabilized.wt, a.c.form, a.cm.form}.
}
\details{
\code{.clean_weights.med()} is called by \code{get_weights.med}, \code{estimate_wtd}, \code{estimate_psYpredMR}, \code{estimate_YpredMR}, \code{estimate_Y2predR}, \code{estimate_MsimYpredMR}.

\code{.clean.weights.te()} is called by \code{get_weights.te} and \code{estimate_psYpred}.

\code{.clean.weights.Ypred()} is called by \code{get_weights.Ypred} and \code{estimate_Ypred}.
}
