% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C0-estimate_wtd.R, R/C1-estimate_psYpred.R,
%   R/C2-estimate_psYpredMR.R, R/C3-estimate_Ypred.R, R/C4-estimate_YpredMR.R,
%   R/C5-estimate_Y2pred.R, R/C6-estimate_Y2predR.R
\name{.clean_y.var}
\alias{.clean_y.var}
\alias{.clean_y.forms.psYpred}
\alias{.clean_y.forms.psYpredMR}
\alias{.clean_y.forms.Ypred}
\alias{.clean_y.forms.YpredMR}
\alias{.clean_y.forms.Y2pred}
\alias{.clean_y.forms.Y2predR}
\alias{dot-clean_y}
\title{Clean outcome-related inputs}
\usage{
.clean_y.var(env)

.clean_y.forms.psYpred(env)

.clean_y.forms.psYpredMR(env)

.clean_y.forms.Ypred(env)

.clean_y.forms.YpredMR(env)

.clean_y.forms.Y2pred(env)

.clean_y.forms.Y2predR(env)
}
\arguments{
\item{env}{An environment}
}
\description{
Internal functions called by \code{estimate_} functions to clean inputs related to the outcome including \code{y.var}, \code{y.c.form}, \code{y.cm.form}, etc.
}
