% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A1-weights_med.R, R/A2-weights_ipw.R,
%   R/A3-weights_cmodds.R, R/C2a-estimate_psYpred.R
\name{.compute_weights.med}
\alias{.compute_weights.med}
\alias{.compute_weights.ipw}
\alias{.compute_weights.cm.odds}
\alias{.compute_weights.psYpred}
\alias{dot-compute_weights}
\title{(For maintainer) Compute weights}
\usage{
.compute_weights.med(data, cross.world, a.c.form, a.cm.form, max.stabilized.wt)

.compute_weights.ipw(data, a.c.form, max.stabilized.wt)

.compute_weights.cm.odds(data, cross.world, a.cm.form, max.stabilized.wt)

.compute_weights.psYpred(data, cross.world, a.c.form, max.stabilized.wt)
}
\arguments{
\item{data}{Prepared internal dataset (e.g., result of \code{.pred_data()}).}

\item{cross.world}{Cleaned cross.world (e.g., from \code{.clean_cross.world()}).}

\item{a.c.form, a.cm.form}{Cleaned formulas (e.g., from \code{.clean_a.forms()}).}

\item{max.stabilized.wt}{A scalar.}
}
\value{
A data frame for the pseudo samples with estimated weights, in stacked format.
}
\description{
Internal functions that compute weights using cleaned inputs.
}
\keyword{internal}
