% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B3-generic_misc.R
\name{.current_RNG}
\alias{.current_RNG}
\alias{.restore_RNG}
\title{A pair of functions to protect RNG state}
\usage{
.current_RNG()

.restore_RNG(RNGstate)
}
\arguments{
\item{RNGstate}{A RNG state, e.g., the output of a previous \code{get_cur_RNG()} call.}
}
\description{
Includes \code{.current_RNG()} which obtains current RNG state, and \code{.restore_RNG()} which restores to a previous RNG state
}
\section{Functions}{
\itemize{
\item \code{.restore_RNG}: 
}}

