% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B2-generic_bootstrap.R
\name{.get_boot.wts}
\alias{.get_boot.wts}
\title{Get weights for bootstrapping.}
\usage{
.get_boot.wts(s.weights, stratify, boot.num, seed, method)
}
\arguments{
\item{s.weights}{Vector of sampling weights. If no sampling weights, use vector of 1s.}

\item{stratify}{Whether to draw bootstrap samples in treated and control units separately.}

\item{boot.num}{How many sets of bootstrap weights (or number of bootstrap samples)}

\item{seed}{Seed for reproducibility}

\item{method}{Bootstrap method. Currently allows "dirichlet", i.e., continuous weights bootstrapping where there weights are drawn from a Dirichlet distribution. TOADD: "multinmial" (or "resample")}
}
\value{
A matrix of bootstrap weights where number of column is number of observations, and each row contains the set of bootstrap weights that forms a bootstrap sample.
}
\description{
Get weights for bootstrapping.
}
