% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cx-shared_estimate.R
\name{.get_contrast}
\alias{.get_contrast}
\title{Contrast two potential outcome means}
\usage{
.get_contrast(a, b, type)
}
\arguments{
\item{a}{The mean of the first potential outcome.}

\item{b}{The mean of the second potential outcome.}

\item{type}{The type of contrast. Options allowed: "MD", "MR", "OR".}
}
\value{
The contrast value.
}
\description{
Contrast two potential outcome means
}
