% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B1-generic_dispatch.R
\name{.get_fun_arg.names}
\alias{.get_fun_arg.names}
\title{Get required argument names from a function}
\usage{
.get_fun_arg.names(fun.name)
}
\arguments{
\item{fun.name}{Name of a function.}
}
\value{
A vector of names of the required arguments of the function
}
\description{
Get required argument names from a function
}
