% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cx-shared_estimate.R
\name{.get_means.and.effects}
\alias{.get_means.and.effects}
\title{Compute potential outcome means from pseudo samples}
\usage{
.get_means.and.effects(w.dat, effect.scale)
}
\arguments{
\item{w.dat}{Pseudo sample data}

\item{effect.scale}{The contrast type}
}
\value{
A named vector of potential outcome means and effects
}
\description{
Compute potential outcome means from pseudo samples
}
