% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B3-generic_misc.R
\name{.get_sd.pooled}
\alias{.get_sd.pooled}
\title{Compute pooled SD}
\usage{
.get_sd.pooled(variable, dat1, dat0)
}
\arguments{
\item{variable}{Name of the variable (covariate or mediator) for which to compute pooled SD.}

\item{dat1, dat0}{Treated and control subsamples.}
}
\value{
Pooled SD.
}
\description{
Compute pooled SD for a variable.
}
