% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B3-generic_misc.R
\name{.get_seed}
\alias{.get_seed}
\title{Sample a seed or seed vector}
\usage{
.get_seed(size = 1, master = NULL)
}
\arguments{
\item{size}{Number of seeds wanted.}

\item{master}{Master seed for reproducibility}
}
\value{
A vector of sampled integers (or a single integer if  \code{size=1}).
}
\description{
This function obtains a vector of integers (or a single integer) to be used as seeds. It preserves RNG state while doing so.
}
