% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B1-generic_dispatch.R
\name{.grab_args}
\alias{.grab_args}
\title{Filter a list of objects for those that are required arguments for an estimator}
\usage{
.grab_args(estimator, arg.list)
}
\arguments{
\item{estimator}{Name of estimator.}

\item{arg.list}{A list of objects containing (but not necessarily limited to) the inputs needed for the estimator.}
}
\description{
Filter a list of objects for those that are required arguments for an estimator
}
