% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A1-weights_med.R, R/A2-weights_ipw.R,
%   R/A3-weights_cmodds.R, R/C2a-estimate_psYpred.R, R/C2b-estimate_psYpredMR.R,
%   R/C3b-estimate_YpredMR.R
\name{.plot_balance.med}
\alias{.plot_balance.med}
\alias{.plot_balance.ipw}
\alias{.plot_balance.odds}
\alias{.plot_balance.psYpred}
\alias{.plot_balance.psYpredMR}
\alias{.plot_balance.YpredMR}
\alias{dot-plot_balance}
\title{(For maintainer) Plot balance}
\usage{
.plot_balance.med(w.dat, c.vars, m.vars, c.std, m.std, key.balance = FALSE)

.plot_balance.ipw(w.dat, vars, vars.std)

.plot_balance.odds(w.dat, vars, vars.std)

.plot_balance.psYpred(w.dat, c.vars, c.std)

.plot_balance.psYpredMR(w.dat, cross.world, c.vars, m.vars, c.std, m.std)

.plot_balance.YpredMR(w.dat, cross.world, c.vars, m.vars, c.std, m.std)
}
\arguments{
\item{w.dat}{Data for (pseudo) subsamples, e.g. \code{output of .compute_weights.Ypred()}}

\item{c.vars}{Names of covariates, already cleaned.}

\item{m.vars}{Names of mediators, already cleaned.}

\item{c.std}{Names of covariates whose mean differences are to be standardized, already cleaned.}

\item{m.std}{Names of mediators whose mean differences are to be standardized, already cleaned.}

\item{key.balance}{Whether all balance components are crucial to the estimator using the weighting method, defaults to FALSE. See relevance in \bold{Value} section.}

\item{vars}{Names of variables whose balance is to be plotted, already cleaned.}

\item{vars.std}{Names of variables whose mean differences are to be standardized, already cleaned.}

\item{cross.world}{(For \code{plot_balance.psYpredMR}) blah}
}
\value{
Plot of balance on the means of covariates and mediators between pseudo subsample(s) and relevant subsample(s).

Plot of balance on the means of covariates and mediators between relevant pseudo samples and full sample. If \code{estimate_wtd==TRUE}, add "(anchor)" and "(for <effect>)" notes to plot labels to draw attention to how each balance matters to the estimator.
}
\description{
Internal functions that makes balance plots for weighted data.
}
\keyword{internal}
