% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A1-weights_med.R, R/A2-weights_ipw.R,
%   R/A3-weights_cmodds.R, R/C2a-estimate_psYpred.R, R/C2b-estimate_psYpredMR.R,
%   R/C3b-estimate_YpredMR.R
\name{.plot_med}
\alias{.plot_med}
\alias{.plot_ipw}
\alias{.plot_cm.odds}
\alias{.plot_psYpred}
\alias{.plot_psYpredMR}
\alias{.plot_YpredMR}
\alias{dot-plot_w.dat}
\title{Plot weight distributions and mean balance}
\usage{
.plot_med(w.dat, c.vars, m.vars, c.vars.std, m.vars.std, estimate.wtd = FALSE)

.plot_ipw(w.dat, c.vars, c.vars.std)

.plot_cm.odds(w.dat, cm.vars, cm.vars.std)

.plot_psYpred(w.dat, c.vars, c.vars.std)

.plot_psYpredMR(w.dat, cross.world, c.vars, m.vars, c.vars.std, m.vars.std)

.plot_YpredMR(w.dat, cross.world, c.vars, m.vars, c.vars.std, m.vars.std)
}
\arguments{
\item{w.dat}{Weighted data}

\item{c.vars}{Names of covariates, already cleaned.}

\item{m.vars}{Names of mediators, already cleaned.}

\item{c.vars.std}{Names of covariates whose mean differences are to be standardized, already cleaned.}

\item{m.vars.std}{Names of mediators whose mean differences are to be standardized, already cleaned.}

\item{estimate.wtd}{Whether plot is part of output of implementing estimator wtd (the pure weighting estimator), defaults to FALSE. See relevance in \bold{Value} section.}

\item{cm.vars}{Names of covariates and mediators.}

\item{cm.vars.std}{Names of covariates and mediators for which to use standardized mean differences in balance plots.}

\item{cross.world}{blah}
}
\description{
Plot weight distributions and mean balance
}
