% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A1-weights_med.R, R/A2-weights_ipw.R,
%   R/A3-weights_cmodds.R, R/C1-estimate_wtd.R, R/C2a-estimate_Y2pred.R,
%   R/C2b-estimate_Y2predR.R, R/C3a-estimate_psYpred.R,
%   R/C3b-estimate_psYpredMR.R, R/C4a-estimate_Ypred.R,
%   R/C4b-estimate_YpredMR.R, R/C5a-estimate_MsimYpred.R,
%   R/D1a-estimate_NDEpred.R, R/D1b-estimate_NDEpredR.R, R/E1-estimate_wtCadj.R,
%   R/E2a-estimate_wpCadj.R, R/E2b-estimate_wpMRCadj.R
\name{.prep_med}
\alias{.prep_med}
\alias{.prep_ipw}
\alias{.prep_odds}
\alias{.prep_wtd}
\alias{.prep_Y2pred}
\alias{.prep_Y2predR}
\alias{.prep_psYpred}
\alias{.prep_psYpredMR}
\alias{.prep_Ypred}
\alias{.prep_YpredMR}
\alias{.prep_MsimYpred}
\alias{.prep_NDEpred}
\alias{.prep_NDEpredR}
\alias{.prep_wtCadj}
\alias{.prep_wpCadj}
\alias{.prep_wpMRCadj}
\alias{dot-prep}
\title{Internal functions: preparing inputs before main computing}
\usage{
.prep_med()

.prep_ipw()

.prep_odds()

.prep_wtd()

.prep_Y2pred()

.prep_Y2predR()

.prep_psYpred()

.prep_psYpredMR()

.prep_Ypred()

.prep_YpredMR()

.prep_MsimYpred()

.prep_NDEpred()

.prep_NDEpredR()

.prep_wtCadj()

.prep_wpCadj()

.prep_wpMRCadj()
}
\description{
A set of internal functions to do the prep work for the main functions before the key computing takes place.
}
\keyword{internal}
