% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B3-generic_misc.R
\name{.reshape_gather}
\alias{.reshape_gather}
\title{A cross between \code{gather} of package \code{dplyr} and \code{reshape}}
\usage{
.reshape_gather(data, columns, key, value, wide.row = TRUE)
}
\arguments{
\item{data}{A data frame}

\item{columns}{Columns to be gathered}

\item{key}{New variable to store names of columns being gathered}

\item{value}{New variable to contain values of columns being gathered}

\item{wide.row}{Whether to include a variable named \code{wide.row} containing the row numbers from the original wide data. Defaults to TRUE.}
}
\value{
A data frame with the two new variables named in arguments \code{key} and \code{value}, plus a third new variable named \code{wide.row}. The remaining variables of the input dataset are carried over.
}
\description{
This function is like \code{gather} and \code{reshape(direction="long")}. Like \code{gather}, it outputs names of variables that are gathered (ie turned from wide to long format). (\code{reshape} outputs a numeric variable for assumed discrete time.) In addition to variable name and value, it outputs (like \code{reshape}) a third variable retaining the wide format row information. \code{reshape} names this variable \code{id}; the current function names the variable \code{wide.row}.
}
\keyword{internal}
