% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C5a-estimate_MsimYpred.R
\name{.simulate}
\alias{.simulate}
\title{Internal: simulate based on a model fit}
\usage{
.simulate(object, ...)
}
\description{
Similar to \code{base::simulate()}, but allows a \code{newdata} argument, similar to that of \code{predict()}.
}
\keyword{internal}
