% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B3-generic_misc.R
\name{.trunc_right}
\alias{.trunc_right}
\title{Right truncation
Right truncation}
\usage{
.trunc_right(vec, max)
}
\arguments{
\item{vec}{A numeric vector.}

\item{max}{A single value of a numeric vector of the same size as \code{vec} indicating the max value(s) used to right-truncate \code{vec}.}
}
\value{
Vector of truncated values.
}
\description{
Right truncation
Right truncation
}
