% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B3-generic_misc.R
\name{dot-wtd_}
\alias{dot-wtd_}
\alias{.wtd_mean}
\alias{.wtd_var}
\title{Weighted statistics}
\usage{
.wtd_mean(x, w)

.wtd_var(x, w)
}
\arguments{
\item{x}{A numeric vector}

\item{w}{A vector of weights}
}
\value{
\code{.wtd_mean()} returns weighted mean. \code{.wtd_var()} returns weighted variance.
}
\description{
Weighted statistics
}
