% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B1-generic_dispatch.R
\name{estimate_effects}
\alias{estimate_effects}
\title{Estimate natural (in)direct effects using an estimator of choice}
\usage{
estimate_effects(
  data,
  s.wt.var = NULL,
  estimator = "Y2predR",
  cross.world = "10",
  effect.scale = "MD",
  boot.num = 999,
  boot.seed = NULL,
  boot.method = "cont-wt",
  boot.stratify = TRUE,
  ...
)
}
\arguments{
\item{data}{A data frame.}

\item{s.wt.var}{Optional, name of variable containing sampling weights.}

\item{estimator}{The estimator of choice. See the paper. -> TOADD citation. Defaults to "Y2pred".}

\item{cross.world}{The cross-world condition involved in the effect decomposition of choice. Should be "10" if want the (NDE0, NIE1) pair, "01" if want the (NIE0, NDE1) pair, or "both" if want both decompositions.}

\item{effect.scale}{The scale of effect of choice. Defaults to "MD" (i.e., mean/risk difference or additive). If outcome is non-negative, also allows "mean ratio" (which could also be specified as "ratio", "MR", "risk ratio", "rate ratio", "RR"). If outcome is binary or bounded within the (0,1) interval, also allows "odds ratio" (which could also be specified as "OR").}

\item{boot.num}{Number of bootstrap samples used for interval estimation, defaults to 999. If just want point estimate, set to 0.}

\item{boot.seed}{Optional, specify bootstrap seed for reproducibility.}

\item{boot.method}{Method for drawing bootstrap samples. Options: "cont-wt" for continuous weights bootstrap, and "resample" for bootstrap by simple resampling (i.e., integer weights bootstrap).}

\item{boot.stratify}{Whether bootstrap samples are drawn stratified by treatment variable. Defaults to TRUE.}

\item{...}{Inputs specific to the estimator.}
}
\value{
A list of objects, including\itemize{
\item{}{}
}
}
\description{
A wrapper function that calls on estimator-specific functions.
}
\keyword{internal}
