% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C5a-estimate_MsimYpred.R
\name{estimate_MsimYpred}
\alias{estimate_MsimYpred}
\title{Estimator MsimYpred}
\usage{
estimate_MsimYpred(
  data,
  s.wt.var = NULL,
  cross.world = "10",
  effect.scale = "additive",
  a.var,
  m.c1.form = NULL,
  m.c0.form = NULL,
  m.c.form = NULL,
  m.dist = NULL,
  y.c1.form = NULL,
  y.c0.form = NULL,
  y.c.form = NULL,
  y.cm1.form = NULL,
  y.cm0.form = NULL,
  y.cm.form = NULL,
  y.link = "identity",
  boot.num = 999,
  boot.seed = NULL,
  boot.method = "cont-wt",
  boot.stratify = TRUE
)
}
\arguments{
\item{data}{A data frame.}

\item{s.wt.var}{Optional, name of variable containing sampling weights.}

\item{cross.world}{The cross-world condition involved in the effect decomposition of choice. Should be "10" if want the (NDE0, NIE1) pair, "01" if want the (NIE0, NDE1) pair, or "both" if want both decompositions.}

\item{effect.scale}{The scale of effect of choice. Defaults to "MD" (i.e., mean/risk difference or additive). If outcome is non-negative, also allows "mean ratio" (which could also be specified as "ratio", "MR", "risk ratio", "rate ratio", "RR"). If outcome is binary or bounded within the (0,1) interval, also allows "odds ratio" (which could also be specified as "OR").}

\item{a.var}{Name of treatment variable.}

\item{m.c1.form, m.c0.form}{blah}

\item{m.c.form}{blah}

\item{m.dist}{blah}

\item{y.c1.form}{blah}

\item{y.c0.form}{blah}

\item{y.c.form}{blah}

\item{y.cm1.form}{blah}

\item{y.cm0.form}{blah}

\item{y.cm.form}{blah}

\item{y.link}{blah}

\item{boot.num}{Number of bootstrap samples used for interval estimation, defaults to 999. If just want point estimate, set to 0.}

\item{boot.seed}{Optional, specify bootstrap seed for reproducibility.}

\item{boot.method}{Method for drawing bootstrap samples. Options: "cont-wt" for continuous weights bootstrap, and "resample" for bootstrap by simple resampling (i.e., integer weights bootstrap).}

\item{boot.stratify}{Whether bootstrap samples are drawn stratified by treatment variable. Defaults to TRUE.}
}
\description{
Function that implements estimator MsimYpred
}
