% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D1a-NDEpred.R
\name{estimate_NDEpred}
\alias{estimate_NDEpred}
\title{Estimator NDEpred}
\usage{
estimate_NDEpred(
  data,
  s.wt.var = NULL,
  cross.world = "10",
  boot.num = 999,
  boot.seed = NULL,
  boot.method = "cont-wt",
  boot.stratify = TRUE,
  a.var,
  y.cm1.form = NULL,
  y.cm0.form = NULL,
  y.cm.form = NULL,
  y.c1.form = NULL,
  y.c0.form = NULL,
  y.c.form = NULL,
  nde0.c.form = NULL,
  nde1.c.form = NULL,
  nde.c.form = NULL,
  y.link = "identity"
)
}
\arguments{
\item{data}{A data frame.}

\item{s.wt.var}{Optional, name of variable containing sampling weights.}

\item{cross.world}{The cross-world condition involved in the effect decomposition of choice. Should be "10" if want the (NDE0, NIE1) pair, "01" if want the (NIE0, NDE1) pair, or "both" if want both decompositions.}

\item{boot.num}{Number of bootstrap samples used for interval estimation, defaults to 999. If just want point estimate, set to 0.}

\item{boot.seed}{Optional, specify bootstrap seed for reproducibility.}

\item{boot.method}{Method for drawing bootstrap samples. Options: "cont-wt" for continuous weights bootstrap, and "resample" for bootstrap by simple resampling (i.e., integer weights bootstrap).}

\item{boot.stratify}{Whether bootstrap samples are drawn stratified by treatment variable. Defaults to TRUE.}

\item{a.var}{Name of treatment variable.}

\item{y.cm1.form}{blah}

\item{y.cm0.form}{blah}

\item{y.cm.form}{blah}

\item{y.c1.form}{blah}

\item{y.c0.form}{blah}

\item{y.c.form}{blah}

\item{nde0.c.form}{Model formula for E[NDE0|C]. Use any name for the response model in this formula, e.g., "nde.prox" or "effect".}

\item{nde1.c.form}{Model formula for E[NDE1|C]. Use any name for response model in this formula, e.g., "nde.prox" or "effect".}

\item{nde.c.form}{Shortcut to specify the same formula for both E[NDE0|C] and E[NDE1|C] models.}

\item{y.link}{blah}
}
\description{
Function that implements estimator NDEYpred
}
\seealso{
Other estimators: 
\code{\link{estimate_Y2predR}()},
\code{\link{estimate_Y2pred}()},
\code{\link{estimate_YpredMR}()},
\code{\link{estimate_Ypred}()},
\code{\link{estimate_psYpredMR}()},
\code{\link{estimate_psYpred}()},
\code{\link{estimate_wtd}()}
}
\concept{additive-effect estimators}
\concept{estimators}
