% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D1b-NDEpredR.R
\name{estimate_NDEpredR}
\alias{estimate_NDEpredR}
\title{Estimator NDEpredR}
\usage{
estimate_NDEpredR(
  data,
  s.wt.var = NULL,
  cross.world = "10",
  boot.num = 999,
  boot.seed = NULL,
  boot.method = "cont-wt",
  boot.stratify = TRUE,
  a.c.form,
  a.cm.form,
  max.stabilized.wt = 30,
  plot = TRUE,
  c.vars.std = NULL,
  m.vars.std = NULL,
  c.order = NULL,
  m.order = NULL,
  y.c1.form = NULL,
  y.c0.form = NULL,
  y.cm1.form = NULL,
  y.cm0.form = NULL,
  nde0.c.form = NULL,
  nde1.c.form = NULL,
  y.c.form = NULL,
  y.cm.form = NULL,
  nde.c.form = NULL,
  y.link = "identity"
)
}
\arguments{
\item{data}{A data frame.}

\item{s.wt.var}{Optional, name of variable containing sampling weights.}

\item{cross.world}{The cross-world condition involved in the effect decomposition of choice. Should be "10" if want the (NDE0, NIE1) pair, "01" if want the (NIE0, NDE1) pair, or "both" if want both decompositions.}

\item{boot.num}{Number of bootstrap samples used for interval estimation, defaults to 999. If just want point estimate, set to 0.}

\item{boot.seed}{Optional, specify bootstrap seed for reproducibility.}

\item{boot.method}{Method for drawing bootstrap samples. Options: "cont-wt" for continuous weights bootstrap, and "resample" for bootstrap by simple resampling (i.e., integer weights bootstrap).}

\item{boot.stratify}{Whether bootstrap samples are drawn stratified by treatment variable. Defaults to TRUE.}

\item{a.c.form}{Formula for the P(A|C) model (the propensity score model).}

\item{a.cm.form}{Formula for the P(A|C,M) model.}

\item{max.stabilized.wt}{Max stabilized weight allowed. Larger weights are truncated to this level. Default is 30.}

\item{plot}{Whether to output weight distribution and balance plots. Defaults to TRUE.}

\item{c.vars.std}{Covariates whose mean differences are to be standardized in balance plot. Ignore if \code{plot==FALSE}.}

\item{m.vars.std}{Mediators whose mean differences are to be standardized in balance plot. Ignore if \code{plot==FALSE}.}

\item{c.order}{Order in which covariates are to be plotted. If not specify, use the order that appears in \code{a.c.form}.}

\item{m.order}{Order in which mediators are to be plotted. If not specify, use the order that appears in \code{a.cm.form}.}

\item{y.c1.form}{blah}

\item{y.c0.form}{blah}

\item{y.cm1.form}{blah}

\item{y.cm0.form}{blah}

\item{nde0.c.form}{Model formula for E[NDE0|C]. Use any name for the response model in this formula, e.g., "nde.prox" or "effect".}

\item{nde1.c.form}{Model formula for E[NDE1|C]. Use any name for response model in this formula, e.g., "nde.prox" or "effect".}

\item{y.c.form}{blah}

\item{y.cm.form}{blah}

\item{nde.c.form}{Shortcut to specify the same formula for both E[NDE0|C] and E[NDE1|C] models.}

\item{y.link}{blah}
}
\description{
Function that implements estimator NDEYpred.R
}
\seealso{
Other estimators: 
\code{\link{estimate_NDEpred}()},
\code{\link{estimate_Y2predR}()},
\code{\link{estimate_Y2pred}()},
\code{\link{estimate_YpredMR}()},
\code{\link{estimate_Ypred}()},
\code{\link{estimate_psYpredMR}()},
\code{\link{estimate_psYpred}()},
\code{\link{estimate_wtd}()}

Other additive-effect estimators: 
\code{\link{estimate_NDEpred}()}

Other more-robust estimators: 
\code{\link{estimate_Y2predR}()}
}
\concept{additive-effect estimators}
\concept{estimators}
\concept{more-robust estimators}
