% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C0-estimate_wtd.R
\name{estimate_wtd}
\alias{estimate_wtd}
\title{Pure weighting estimator}
\usage{
estimate_wtd(
  data,
  s.wt.var = NULL,
  cross.world = "10",
  effect.scale = "MD",
  boot.num = 999,
  boot.seed = NULL,
  boot.method = "cont-wt",
  boot.stratify = TRUE,
  a.c.form,
  a.cm.form,
  y.var,
  max.stabilized.wt = 30,
  plot = TRUE,
  c.vars.std = NULL,
  m.vars.std = NULL,
  c.order = NULL,
  m.order = NULL
)
}
\arguments{
\item{data}{A data frame.}

\item{s.wt.var}{Optional, name of variable containing sampling weights.}

\item{cross.world}{The cross-world condition involved in the effect decomposition of choice. Should be "10" if want the (NDE0, NIE1) pair, "01" if want the (NIE0, NDE1) pair, or "both" if want both decompositions.}

\item{effect.scale}{The scale of effect of choice. Defaults to "MD" (i.e., mean/risk difference or additive). If outcome is non-negative, also allows "mean ratio" (which could also be specified as "ratio", "MR", "risk ratio", "rate ratio", "RR"). If outcome is binary or bounded within the (0,1) interval, also allows "odds ratio" (which could also be specified as "OR").}

\item{boot.num}{Number of bootstrap samples used for interval estimation, defaults to 999. If just want point estimate, set to 0.}

\item{boot.seed}{Optional, specify bootstrap seed for reproducibility.}

\item{boot.method}{Method for drawing bootstrap samples. Options: "cont-wt" for continuous weights bootstrap, and "resample" for bootstrap by simple resampling (i.e., integer weights bootstrap).}

\item{boot.stratify}{Whether bootstrap samples are drawn stratified by treatment variable. Defaults to TRUE.}

\item{a.c.form}{Formula for the P(A|C) model (the propensity score model).}

\item{a.cm.form}{Formula for the P(A|C,M) model.}

\item{y.var}{Name of the outcome variable.}

\item{max.stabilized.wt}{Max stabilized weight allowed. Larger weights are truncated to this level. Default is 30.}

\item{plot}{Whether to output weight distribution and balance plots. Defaults to TRUE.}

\item{c.vars.std}{Covariates whose mean differences are to be standardized in balance plot. Ignore if \code{plot==FALSE}.}

\item{m.vars.std}{Mediators whose mean differences are to be standardized in balance plot. Ignore if \code{plot==FALSE}.}

\item{c.order}{Order in which covariates are to be plotted. If not specify, use the order that appears in \code{a.c.form}.}

\item{m.order}{Order in which mediators are to be plotted. If not specify, use the order that appears in \code{a.cm.form}.}
}
\value{
If \code{plot==FALSE} and \code{boot.num==0}, a point estimate vector including the potential outcome means and the effects.

Otherwise, a list of objects including:\tabular{ll}{
\code{estimates} \tab point estimates (and if \code{boot.num>0}) 95\% quantile intervals and standard errors \cr
\cr
\code{boot.seed} \tab the seed used for the bootstrap (if \code{boot.num>0}) \cr
\cr
\code{wt.dist.plot} \tab a \code{ggplot2} plot of weight distributions (if \code{plot==TRUE}) \cr
\cr
\code{balance.plot} \tab a \code{ggplot2} plot of balance on the means of covariate and mediator for the pseudo samples (if \code{plot==TRUE}) \cr
}
}
\description{
Function that implements the pure weighting estimator of natural (in)direct effects.
}
\seealso{
Other estimators: 
\code{\link{estimate_NDEpred}()},
\code{\link{estimate_Y2predR}()},
\code{\link{estimate_Y2pred}()},
\code{\link{estimate_YpredMR}()},
\code{\link{estimate_Ypred}()},
\code{\link{estimate_psYpredMR}()},
\code{\link{estimate_psYpred}()}
}
\concept{estimators}
