% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C1a-estimate_Y2pred.R
\name{estimate_Y2pred}
\alias{estimate_Y2pred}
\title{Estimator Y2pred}
\usage{
estimate_Y2pred(
  data,
  s.wt.var = NULL,
  cross.world = "10",
  effect.scale = "additive",
  boot.num = 999,
  boot.seed = NULL,
  boot.method = "cont-wt",
  boot.stratify = TRUE,
  a.var,
  y.cm.form = NULL,
  y.cm1.form = NULL,
  y.cm0.form = NULL,
  y.c.form = NULL,
  y.c1.form = NULL,
  y.c0.form = NULL,
  y10.c.form = NULL,
  y01.c.form = NULL,
  y.link = "identity"
)
}
\arguments{
\item{data}{A data frame.}

\item{s.wt.var}{Optional, name of variable containing sampling weights.}

\item{cross.world}{The cross-world condition involved in the effect decomposition of choice. Should be "10" if want the (NDE0, NIE1) pair, "01" if want the (NIE0, NDE1) pair, or "both" if want both decompositions.}

\item{effect.scale}{The scale of effect of choice. Defaults to "MD" (i.e., mean/risk difference or additive). If outcome is non-negative, also allows "mean ratio" (which could also be specified as "ratio", "MR", "risk ratio", "rate ratio", "RR"). If outcome is binary or bounded within the (0,1) interval, also allows "odds ratio" (which could also be specified as "OR").}

\item{boot.num}{Number of bootstrap samples used for interval estimation, defaults to 999. If just want point estimate, set to 0.}

\item{boot.seed}{Optional, specify bootstrap seed for reproducibility.}

\item{boot.method}{Method for drawing bootstrap samples. Options: "cont-wt" for continuous weights bootstrap, and "resample" for bootstrap by simple resampling (i.e., integer weights bootstrap).}

\item{boot.stratify}{Whether bootstrap samples are drawn stratified by treatment variable. Defaults to TRUE.}

\item{a.var}{Name of treatment variable.}

\item{y.cm.form}{blah}

\item{y.cm1.form}{blah}

\item{y.cm0.form}{blah}

\item{y.c.form}{blah}

\item{y.c1.form}{blah}

\item{y.c0.form}{blah}

\item{y10.c.form}{Model formula for E[Y(1,M0)|C]}

\item{y01.c.form}{Model formula for E[Y(0,M1)|C]}

\item{y.link}{blah}
}
\description{
Function that implements estimator Y2pred
}
\seealso{
Other estimators: 
\code{\link{estimate_NDEpredR}()},
\code{\link{estimate_NDEpred}()},
\code{\link{estimate_Y2predR}()},
\code{\link{estimate_YpredMR}()},
\code{\link{estimate_Ypred}()},
\code{\link{estimate_psYpredMR}()},
\code{\link{estimate_psYpred}()},
\code{\link{estimate_wtCadj}()},
\code{\link{estimate_wtd}()}
}
\concept{estimators}
