% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3a-estimate_Ypred.R
\name{estimate_Ypred}
\alias{estimate_Ypred}
\title{Estimator Ypred}
\usage{
estimate_Ypred(
  data,
  s.wt.var = NULL,
  cross.world = "10",
  effect.scale = "MD",
  boot.num = 999,
  boot.seed = NULL,
  boot.method = "cont-wt",
  boot.stratify = TRUE,
  a.cm.form,
  max.stabilized.wt = 30,
  plot = TRUE,
  cm.vars.std = NULL,
  cm.order = NULL,
  y.c.form = NULL,
  y.c1.form = NULL,
  y.c0.form = NULL,
  y10.c.form = NULL,
  y01.c.form = NULL,
  y.link = "identity"
)
}
\arguments{
\item{data}{A data frame.}

\item{s.wt.var}{Optional, name of variable containing sampling weights.}

\item{cross.world}{The cross-world condition involved in the effect decomposition of choice. Should be "10" if want the (NDE0, NIE1) pair, "01" if want the (NIE0, NDE1) pair, or "both" if want both decompositions.}

\item{effect.scale}{The scale of effect of choice. Defaults to "MD" (i.e., mean/risk difference or additive). If outcome is non-negative, also allows "mean ratio" (which could also be specified as "ratio", "MR", "risk ratio", "rate ratio", "RR"). If outcome is binary or bounded within the (0,1) interval, also allows "odds ratio" (which could also be specified as "OR").}

\item{boot.num}{Number of bootstrap samples used for interval estimation, defaults to 999. If just want point estimate, set to 0.}

\item{boot.seed}{Optional, specify bootstrap seed for reproducibility.}

\item{boot.method}{Method for drawing bootstrap samples. Options: "cont-wt" for continuous weights bootstrap, and "resample" for bootstrap by simple resampling (i.e., integer weights bootstrap).}

\item{boot.stratify}{Whether bootstrap samples are drawn stratified by treatment variable. Defaults to TRUE.}

\item{a.cm.form}{Formula for the P(A|C,M) model.}

\item{max.stabilized.wt}{Level at which weights (on stabilized scale) should be right-truncated. Default to 30, meaning no unit should count more than as 30 persons.}

\item{plot}{Whether to output weight distribution and balance plots. Defaults to TRUE.}

\item{cm.vars.std}{Names of covariates and mediators whose mean differences are to be standardized in balance plotting. Ignore if \code{plot==FALSE}.}

\item{cm.order}{Order in which covariates and mediators are to be plotted. If not specify, use order that appears in \code{a.cm.form}.}

\item{y.c.form}{blah}

\item{y.c1.form}{blah}

\item{y.c0.form}{blah}

\item{y10.c.form}{Model formula for E[Y(1,M0)|C]}

\item{y01.c.form}{Model formula for E[Y(0,M1)|C]}

\item{y.link}{blah}
}
\description{
Function that implements estimator Ypred
}
\seealso{
[estimate_YpredMR()] which is closely related but more robust

Other estimators: 
\code{\link{estimate_Y2predR}()},
\code{\link{estimate_Y2pred}()},
\code{\link{estimate_YpredMR}()},
\code{\link{estimate_psYpredMR}()},
\code{\link{estimate_psYpred}()},
\code{\link{estimate_wtd}()}
}
\concept{estimators}
