% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediationClarity.R
\docType{package}
\name{mediationClarity-package}
\alias{mediationClarity}
\alias{mediationClarity-package}
\title{mediationClarity: Simple and more robust estimation of marginal natural (in)direct effects}
\description{
This package implements estimators covered in Nguyen et al. (2022), arXiv:2102.06048.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/trangnguyen74/mediationClarity}
  \item Report bugs at \url{https://github.com/trangnguyen74/mediationClarity/issues}
}

}
\keyword{internal}
