% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ax-shared_weights.R
\name{stabilize_weight}
\alias{stabilize_weight}
\title{Standardize a weight variable by group}
\usage{
stabilize_weight(weight, group = NULL, s.weight = NULL)
}
\arguments{
\item{weight}{A weight vector to be standardized}

\item{group}{Optional. A character vector indicating group membership.}

\item{s.weight}{Optional. A vector of sampling weight.}
}
\description{
Standardize a weight variable by group
}
