% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B3-generic_misc.R
\docType{data}
\name{synth}
\alias{synth}
\title{This is data to be included in my package}
\description{
This is data to be included in my package
}
\references{
\url{data_blah.com}
}
\author{
My Name \email{blahblah@roxygen.org}
}
\keyword{data}
