% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnMediation-package.R
\docType{package}
\name{tnMediation-package}
\alias{tnMediation}
\alias{tnMediation-package}
\title{tnMediation: Simple and more robust estimation of marginal natural (in)direct effects}
\description{
This package implements estimators covered in Nguyen et al. (2022), arXiv:2102.06048.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/trangnguyen74/tnMediation}
  \item Report bugs at \url{https://github.com/trangnguyen74/tnMediation/issues}
}

}
\keyword{internal}
