% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A3-weights_cmodds.R
\name{weights_cm.odds}
\alias{weights_cm.odds}
\title{Estimates odds weights for the Ypred estimator}
\usage{
weights_cm.odds(
  data,
  s.wt.var = NULL,
  cross.world = "10",
  a.cm.form,
  max.stabilized.wt = 30,
  plot = TRUE,
  cm.order = NULL,
  cm.vars.std = NULL
)
}
\arguments{
\item{data}{A data frame.}

\item{s.wt.var}{Optional, name of variable containing sampling weights.}

\item{cross.world}{The cross-world condition involved in the effect decomposition of choice. Should be "10" if want the (NDE0, NIE1) pair, "01" if want the (NIE0, NDE1) pair, or "both" if want both decompositions.}

\item{a.cm.form}{Formula for the P(A|C,M) model.}

\item{max.stabilized.wt}{Level at which weights (on stabilized scale) should be right-truncated. Default to 30, meaning no unit should count more than as 30 persons.}

\item{plot}{Whether to output weight distribution and balance plots. Defaults to TRUE.}

\item{cm.order}{Order in which covariates and mediators are to be plotted. If not specify, use order that appears in \code{a.cm.form}.}

\item{cm.vars.std}{Names of covariates and mediators whose mean differences are to be standardized in balance plotting. Ignore if \code{plot==FALSE}.}
}
\value{
\tabular{ll}{
\code{w.dat} \tab A data frame for the pseudo and actual subsamples (pseudo ones with estimated distribution-morphing odds weights) \cr
\code{wt.dist.plot} \tab If \code{plot==TRUE}, a density plot of the distribution-morphing weights, and (if sampling weights vary) a density plot of the final weights \cr
\code{balance.plot} \tab If \code{plot==TRUE}, a plot of the balance in covariates and mediators of the pseudo subsample(s).
}
}
\description{
Estimates odds weights for the Ypred estimator
}
\seealso{
Other weighting schemes: 
\code{\link{weights_ipw}()},
\code{\link{weights_med}()}
}
\concept{weighting schemes}
