% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A2-weights_ipw.R
\name{weights_ipw}
\alias{weights_ipw}
\title{Estimate inverse probability weights}
\usage{
weights_ipw(
  data,
  s.wt.var = NULL,
  a.c.form,
  max.stabilized.wt = 30,
  plot = TRUE,
  c.order = NULL,
  c.vars.std = NULL
)
}
\arguments{
\item{data}{A data frame.}

\item{s.wt.var}{Optional, name of variable containing sampling weights.}

\item{a.c.form}{Formula for the P(A|C) model (the propensity score model).}

\item{max.stabilized.wt}{Max stabilized weight allowed. Larger weights are truncated to this level. Default is 30.}

\item{plot}{Whether to output weight distribution and balance plots. Defaults to TRUE.}

\item{c.order}{Order in which covariates are to be plotted. If not specify, use the order that appears in \code{a.c.form}.}

\item{c.vars.std}{Covariates whose mean differences are to be standardized in balance plot. Ignore if \code{plot==FALSE}.}
}
\value{
A list including\tabular{ll}{
\code{w.dat} \tab A data frame for the pseudo samples with estimated weights. \cr
\code{plot.wts} \tab A plot of the distributions of the weights. \cr
\code{plot.balance} \tab A plot of the balance in covariates and mediators of the pseudo samples. \cr
}
}
\description{
This function estimates weights that form two pseudo treated and pseudo control samples (as if for total effect estimation)
}
\seealso{
Other weighting schemes: 
\code{\link{weights_cm.odds}()},
\code{\link{weights_med}()}
}
\concept{weighting schemes}
