% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A1-weights_med.R
\name{weights_med}
\alias{weights_med}
\title{Estimate weights for natural (in)direct effects estimation}
\usage{
weights_med(
  data,
  s.wt.var = NULL,
  cross.world = "10",
  a.c.form,
  a.cm.form,
  max.stabilized.wt = 30,
  plot = TRUE,
  c.order = NULL,
  m.order = NULL,
  c.vars.std = NULL,
  m.vars.std = NULL
)
}
\arguments{
\item{data}{A data frame.}

\item{s.wt.var}{Optional, name of variable containing sampling weights.}

\item{cross.world}{The cross-world condition involved in the effect decomposition of choice. Should be "10" if want the (NDE0, NIE1) pair, "01" if want the (NIE0, NDE1) pair, or "both" if want both decompositions.}

\item{a.c.form}{Formula for the P(A|C) model (the propensity score model).}

\item{a.cm.form}{Formula for the P(A|C,M) model.}

\item{max.stabilized.wt}{Level at which weights (on stabilized scale) should be right-truncated. Default to 30, meaning no unit should count more than as 30 persons.}

\item{plot}{Whether to output weight distribution and balance plots. Defaults to TRUE.}

\item{c.order}{Order in which covariates are to be plotted. If not specify, use the order that appears in \code{a.c.form}.}

\item{m.order}{Order in which mediators are to be plotted. If not specify, use the order that appears in \code{a.cm.form}.}

\item{c.vars.std}{Covariates whose mean differences are to be standardized in balance plot. Ignore if \code{plot==FALSE}.}

\item{m.vars.std}{Mediators whose mean differences are to be standardized in balance plot. Ignore if \code{plot==FALSE}.}
}
\value{
A list including\itemize{
\item{w.dat}{A data frame for the pseudo samples with estimated weights.}
\item{plot.wts}{A plot of the distributions of the weights.}
\item{plot.balance}{A plot of the balance in covariates and mediators of the pseudo samples.}
}
}
\description{
Estimates inverse probability weights that form pseudo treated and control samples and cross-world weights that form (one or two) pseudo cross-world samples.
}
\seealso{
Other weighting schemes: 
\code{\link{weights_cm.odds}()},
\code{\link{weights_ipw}()}
}
\concept{weighting schemes}
