% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A3-weights_odds.R
\name{weights_odds}
\alias{weights_odds}
\title{Estimates odds weights for the Ypred estimator}
\usage{
weights_odds(
  data,
  s.wt.var = NULL,
  cross.world = "10",
  a.form,
  max.stabilized.wt = 30,
  plot = TRUE,
  vars.order = NULL,
  vars.std = NULL
)
}
\arguments{
\item{data}{A data frame.}

\item{s.wt.var}{Optional, name of variable containing sampling weights.}

\item{cross.world}{The cross-world condition involved in the effect decomposition of choice. Should be "10" if want the (NDE0, NIE1) pair, "01" if want the (NIE0, NDE1) pair, or "both" if want both decompositions.}

\item{a.form}{Formula for model of treatment given the variables that need to be balanced.}

\item{max.stabilized.wt}{Max stabilized weight allowed. Larger weights are truncated to this level. Default is 30.}

\item{plot}{Whether to output weight distribution and balance plots. Defaults to TRUE.}

\item{vars.order}{Order in which variables are to appear in the balance plot. If not specify, use the order that appears in \code{a.form}.}

\item{vars.std}{Variables whose mean differences are to be standardized in balance plot. Ignore if \code{plot==FALSE}.}
}
\value{
\tabular{ll}{
\code{w.dat} \tab A data frame for the pseudo and actual subsamples (pseudo ones with estimated distribution-morphing odds weights) \cr
\code{wt.dist.plot} \tab If \code{plot==TRUE}, a density plot of the distribution-morphing weights, and (if sampling weights vary) a density plot of the final weights \cr
\code{balance.plot} \tab If \code{plot==TRUE}, a plot of the balance in covariates and mediators of the pseudo subsample(s).
}
}
\description{
Estimates odds weights for the Ypred estimator
}
\seealso{
Other weighting schemes: 
\code{\link{weights_ipw}()},
\code{\link{weights_med}()}
}
\concept{weighting schemes}
