% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A1-weights_med.R, R/A2-weights_ipw.R,
%   R/A3-weights_cmodds.R, R/C3a-estimate_Ypred.R
\name{.check_plot.med}
\alias{.check_plot.med}
\alias{.check_plot.ipw}
\alias{.check_plot.cm.odds}
\alias{.check_plot.Ypred}
\alias{dot-check_plot}
\title{Internal functions: check inputs for plotting}
\usage{
.check_plot.med(env)

.check_plot.ipw(env)

.check_plot.cm.odds(env)

.check_plot.Ypred(env)
}
\arguments{
\item{env}{An environment}
}
\description{
Functions called by \code{.prep_} functions within \code{weights_} an \code{estimate_} functions.
}
\details{
\code{.check_plot.med()} is called by \code{.prep_med()}.

\code{.check_plot.ipw()} is called by \code{.prep_ipw()}.

\code{.check_plot.odds()} is called by \code{.prep_odds()}.

\code{.check_plot.Ypred()} is called by \code{.prep_Ypred()}.
}
\keyword{internal}
