% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A1-weights_med.R, R/A2-weights_ipw.R,
%   R/A3-weights_cmodds.R, R/C2a-estimate_Y2pred.R, R/C3a-estimate_psYpred.R,
%   R/C4a-estimate_Ypred.R
\name{.clean_weights.med}
\alias{.clean_weights.med}
\alias{.clean_weights.ipw}
\alias{.clean_weights.odds}
\alias{.clean_a.var}
\alias{.clean_weights.psYpred}
\alias{.clean_weights.Ypred}
\alias{dot-clean_weights}
\title{Internal functions: clean inputs for weighting}
\usage{
.clean_weights.med(env)

.clean_weights.ipw(env)

.clean_weights.odds(env)

.clean_a.var(env)

.clean_weights.psYpred(env)

.clean_weights.Ypred(env)
}
\arguments{
\item{env}{An environment}
}
\description{
Functions called by \code{.prep_} functions within \code{weights_} and \code{estimate_} functions. If no weighting is required, grab treatment variable.
}
\details{
\code{.clean_weights.med()} is used by \code{.prep_med()}.

\code{.clean_weights.ipw()} is used by \code{.prep_ipw()}.

\code{.clean_weights.odds()} is used by \code{.prep_odds()}.

\code{.clean_a.var()} is used by \code{.prep_Y2pred()}, which is used by \code{estimate_Y2pred()}.

\code{.clean_weights.psYpred()} is used by \code{.prep_psYpred()}.

\code{.clean_weights.Ypred()} is used by \code{.prep_Ypred()}.
}
\keyword{internal}
