% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C0-estimate_wtd.R, R/C2a-estimate_psYpred.R,
%   R/C2b-estimate_psYpredMR.R, R/C3a-estimate_Ypred.R,
%   R/C3b-estimate_YpredMR.R, R/C1a-estimate_Y2pred.R, R/C1b-estimate_Y2predR.R,
%   R/D1a-NDEpred.R, R/E1a-wtCadj.R
\name{.clean_y.wtd}
\alias{.clean_y.wtd}
\alias{.clean_y.forms.psYpred}
\alias{.clean_y.forms.psYpredMR}
\alias{.clean_y.forms.Ypred}
\alias{.clean_y.forms.YpredMR}
\alias{.clean_y.forms.Y2pred}
\alias{.clean_y.forms.Y2predR}
\alias{.clean_ye.forms.NDEpred}
\alias{.clean_y.wtCadj}
\alias{dot-clean_y}
\title{(For maintainer) Clean outcome-related inputs}
\usage{
.clean_y.wtd(env)

.clean_y.forms.psYpred(env)

.clean_y.forms.psYpredMR(env)

.clean_y.forms.Ypred(env)

.clean_y.forms.YpredMR(env)

.clean_y.forms.Y2pred(env)

.clean_y.forms.Y2predR(env)

.clean_ye.forms.NDEpred(env)

.clean_y.wtCadj(env)
}
\arguments{
\item{env}{An environment}
}
\description{
Internal functions called by \code{estimate_} functions to clean inputs related to the outcome including \code{y.var}, \code{y.c.form}, \code{y.cm.form}, etc.
}
\keyword{internal}
