% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A1-weights_med.R, R/A2-weights_ipw.R,
%   R/A3-weights_cmodds.R, R/C1-estimate_psYpred.R, R/C2-estimate_psYpredMR.R,
%   R/C4-estimate_YpredMR.R
\name{.plot_balance.med}
\alias{.plot_balance.med}
\alias{.plot_balance.ipw}
\alias{.plot_balance.cm.odds}
\alias{.plot_balance.psYpred}
\alias{.plot_balance.psYpredMR}
\alias{.plot_balance.YpredMR}
\alias{dot-plot_balance}
\title{Plot balance}
\usage{
.plot_balance.med(
  w.dat,
  c.vars,
  m.vars,
  c.vars.std,
  m.vars.std,
  estimate.wtd = FALSE
)

.plot_balance.ipw(w.dat, c.vars, c.vars.std)

.plot_balance.cm.odds(w.dat, cm.vars, cm.vars.std)

.plot_balance.psYpred(w.dat, c.vars, c.vars.std)

.plot_balance.psYpredMR(
  w.dat,
  cross.world,
  c.vars,
  m.vars,
  c.vars.std,
  m.vars.std
)

.plot_balance.YpredMR(
  w.dat,
  cross.world,
  c.vars,
  m.vars,
  c.vars.std,
  m.vars.std
)
}
\arguments{
\item{w.dat}{Data for (pseudo) subsamples, e.g. \code{output of .compute_weights.Ypred()}}

\item{c.vars}{Names of covariates, already cleaned.}

\item{m.vars}{Names of mediators, already cleaned.}

\item{c.vars.std}{Names of covariates whose mean differences are to be standardized, already cleaned.}

\item{m.vars.std}{Names of mediators whose mean differences are to be standardized, already cleaned.}

\item{estimate.wtd}{Whether plot is part of output of implementing estimator wtd (the pure weighting estimator), defaults to FALSE. See relevance in \bold{Value} section.}

\item{cm.vars}{Names of covariates and mediators, already cleaned.}

\item{cm.vars.std}{Names of mediators whose mean differences are to be standardized, already cleaned}

\item{cross.world}{(For \code{plot_balance.psYpredMR}) blah}
}
\value{
Plot of balance on the means of covariates and mediators between pseudo subsample(s) and relevant subsample(s).

Plot of balance on the means of covariates and mediators between relevant pseudo samples and full sample. If \code{estimate_wtd==TRUE}, add "(anchor)" and "(for <effect>)" notes to plot labels to draw attention to how each balance matters to the estimator.
}
\description{
Internal function that makes balance plot for mediation weights and TE weights.

Internal functions that makes balance plots for weighted data.
}
