% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A1-weights_med.R, R/A2-weights_ipw.R,
%   R/A3-weights_cmodds.R, R/C3a-estimate_psYpred.R, R/C3b-estimate_psYpredMR.R,
%   R/C4b-estimate_YpredMR.R, R/E2b-estimate_wpMRCadj.R
\name{.plot_med}
\alias{.plot_med}
\alias{.plot_ipw}
\alias{.plot_odds}
\alias{.plot_psYpred}
\alias{.plot_psYpredMR}
\alias{.plot_YpredMR}
\alias{.plot_wpMRCadj}
\alias{dot-plot_w.dat}
\title{(For maintainer) Plot weight distributions and mean balance}
\usage{
.plot_med(w.dat, c.vars, m.vars, c.std, m.std, key.balance = FALSE)

.plot_ipw(w.dat, vars, vars.std, key.balance = FALSE)

.plot_odds(w.dat, vars, vars.std, estimate.Ypred = FALSE)

.plot_psYpred(w.dat, c.vars, c.std)

.plot_psYpredMR(w.dat, cross.world, c.vars, m.vars, c.std, m.std)

.plot_YpredMR(w.dat, cross.world, c.vars, m.vars, c.std, m.std)

.plot_wpMRCadj(w.dat, c.vars, m.vars, c.std, m.std)
}
\arguments{
\item{w.dat}{Weighted data}

\item{c.vars}{Names of covariates, already cleaned.}

\item{m.vars}{Names of mediators, already cleaned.}

\item{c.std}{Names of covariates whose mean differences are to be standardized, already cleaned.}

\item{m.std}{Names of mediators whose mean differences are to be standardized, already cleaned.}

\item{key.balance}{Whether all balance components are crucial to the estimator using the weighting method, defaults to FALSE. See relevance in \bold{Value} section.}

\item{vars}{Names of covariates and mediators.}

\item{vars.std}{Names of covariates and mediators for which to use standardized mean differences in balance plots.}

\item{cross.world}{blah}
}
\description{
(For maintainer) Plot weight distributions and mean balance
}
\keyword{internal}
