% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C0-estimate_wtd.R, R/C2a-estimate_psYpred.R,
%   R/C2b-estimate_psYpredMR.R, R/C3a-estimate_Ypred.R,
%   R/C3b-estimate_YpredMR.R, R/C1a-estimate_Y2pred.R, R/C1b-estimate_Y2predR.R,
%   R/D1a-NDEpred.R, R/D1b-NDEpredR.R, R/E1-wtCadj.R
\name{.point_est.wtd}
\alias{.point_est.wtd}
\alias{.point_est.psYpred}
\alias{.point_est.psYpredMR}
\alias{.point_est.Ypred}
\alias{.point_est.YpredMR}
\alias{.point_est.Y2pred}
\alias{.point_est.Y2predR}
\alias{.point_est.NDEpred}
\alias{.point_est.NDEpredR}
\alias{.point_est.wtCadj}
\alias{dot-point_est}
\title{(For maintainer) Point estimation}
\usage{
.point_est.wtd(
  data,
  cross.world,
  effect.scale,
  a.c.form,
  a.cm.form,
  max.stabilized.wt = 30,
  output.data = FALSE
)

.point_est.psYpred(
  data,
  cross.world,
  effect.scale,
  a.c.form,
  max.stabilized.wt = 30,
  output.data = FALSE,
  y.c1.form,
  y.c0.form,
  y.cm1.form,
  y.cm0.form,
  y.family
)

.point_est.psYpredMR(
  data,
  cross.world,
  effect.scale,
  a.c.form,
  a.cm.form,
  max.stabilized.wt = 30,
  output.data = FALSE,
  y.c1.form,
  y.c0.form,
  y.cm1.form,
  y.cm0.form,
  y.family
)

.point_est.Ypred(
  data,
  cross.world,
  effect.scale,
  a.cm.form,
  max.stabilized.wt = 30,
  output.data = FALSE,
  y.c1.form,
  y.c0.form,
  y10.c.form = NULL,
  y01.c.form = NULL,
  y.family
)

.point_est.YpredMR(
  data,
  cross.world,
  effect.scale,
  a.c.form,
  a.cm.form,
  max.stabilized.wt = 30,
  output.data = FALSE,
  y.c1.form,
  y.c0.form,
  y10.c.form = NULL,
  y01.c.form = NULL,
  y.family
)

.point_est.Y2pred(
  data,
  cross.world,
  effect.scale,
  y.c1.form,
  y.c0.form,
  y.cm1.form,
  y.cm0.form,
  y10.c.form,
  y01.c.form,
  y.family,
  output.data = FALSE
)

.point_est.Y2predR(
  data,
  cross.world,
  effect.scale,
  a.c.form,
  a.cm.form,
  max.stabilized.wt = 30,
  output.data = FALSE,
  y.c1.form,
  y.c0.form,
  y.cm1.form,
  y.cm0.form,
  y10.c.form,
  y01.c.form,
  y.family
)

.point_est.NDEpred(
  data,
  cross.world,
  y.c1.form,
  y.c0.form,
  y.cm1.form,
  y.cm0.form,
  nde0.c.form,
  nde1.c.form,
  y.family,
  output.data = FALSE
)

.point_est.NDEpredR(
  data,
  cross.world,
  a.c.form,
  a.cm.form,
  max.stabilized.wt = 30,
  output.data = FALSE,
  y.c1.form,
  y.c0.form,
  y.cm1.form,
  y.cm0.form,
  nde0.c.form,
  nde1.c.form,
  y.family
)

.point_est.wtCadj(
  data,
  cross.world,
  effect.scale,
  a.c.form,
  a.cm.form,
  max.stabilized.wt = 30,
  output.data = FALSE,
  wkng.form,
  y.family
)
}
\arguments{
\item{data}{A dataset that has been prepared to have the sampling with variable named s.wt (e.g., using function \code{.clean_inputs_generic()}.}

\item{cross.world}{Three options: "10", "01" or c("10", "01").}

\item{effect.scale}{Three options: "MD", "MR", "OR".}

\item{a.c.form}{Formula for the P(A|C) model (the propensity score model).}

\item{a.cm.form}{Formula for the P(A|C,M) model.}

\item{max.stabilized.wt}{Max stabilized weight allowed. Larger weights are truncated to this level. Default is 30.}

\item{output.data}{Whether to output the weighted data in addition to the estimated potential outcome means and effects. Defaults to FALSE.}

\item{y.c1.form, y.c0.form}{Model formulas for E[Y|C,A=1], E[Y|C,A=0], checked.}

\item{y.cm1.form, y.cm0.form}{Model formulas for E[Y|C,M,A=1], E[Y|C,M,A=0], checked.}

\item{y.family}{GLM (quasi-) family for outcome models.}

\item{y10.c.form, y01.c.form}{Model formulas for E[Y(1,M0)|C] and E[Y(0,M1)|C].}

\item{wkng.form}{Formula for working regression model}
}
\value{
If \code{plot==FALSE}, effect estimates (numeric vector). If \code{plot==TRUE}, a list including this vector of effect estimates and the weight distribution and mean balance plots.
}
\description{
Internal functions to be called within \code{estimate_\*} to obtain the point estimate of effects.
}
\keyword{internal}
