% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A1-weights_med.R, R/A2-weights_ipw.R,
%   R/A3-weights_cmodds.R, R/C0-estimate_wtd.R, R/C2a-estimate_psYpred.R,
%   R/C2b-estimate_psYpredMR.R, R/C3a-estimate_Ypred.R,
%   R/C3b-estimate_YpredMR.R, R/C1a-estimate_Y2pred.R, R/C1b-estimate_Y2predR.R
\name{.prep_med}
\alias{.prep_med}
\alias{.prep_ipw}
\alias{.prep_cm.odds}
\alias{.prep_estimate_wtd}
\alias{.prep_psYpred}
\alias{.prep_psYpredMR}
\alias{.prep_Ypred}
\alias{.prep_YpredMR}
\alias{.prep_Y2pred}
\alias{.prep_Y2predR}
\alias{dot-prep}
\title{Prep work within main functions}
\usage{
.prep_med()

.prep_ipw()

.prep_cm.odds()

.prep_estimate_wtd()

.prep_psYpred()

.prep_psYpredMR()

.prep_Ypred()

.prep_YpredMR()

.prep_Y2pred()

.prep_Y2predR()
}
\description{
A set of internal functions to do the prep work for the main functions before the key computing takes place.
}
